API-02 Mudança no tipo do Id esperado pelo repositorio paciente

+++ src/main/java/com/takecare/takecareapi/repository/PacienteRepository.java
@@ -6,7 +6,7 @@ import com.takecare.takecareapi.entities.Paciente;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PacienteRepository extends JpaRepository<Paciente, Long> {
+public interface PacienteRepository extends JpaRepository<Paciente, Integer> {
     List<Paciente> findAll();
 
     Paciente findById(int Id);
